Added Cryptography.c/.h and modified the existing hashing.c/.h to include additional functions, other functions should work as expected. Test.c modified to include examples for the new crypto functions. For details please see the below...
**NOTE: cryptography.h includes hashing.h which may cause C compiler problems if hashing.h is already included in the main program**

crypto struct: IV -> initialization vector, random per file and per save, functions like a salt // ciphertext -> the encrypted text // ct_len -> the length of the ciphertext, padded to a full block at the end
AES_KEY is currently fixed for the program like a pepper is
addPadding and removePadding [internal function] : due to some bugs in the openSSL padding scheme, I implemented my own which uses the remaining bytes as the pad, so 5 bytes to a full block is 55555 at the end
genRandBytes [internal function] : generates a random string like the IV
encryptString : AES-cbc 256 encrypt a string and return the crypto struct
decryptString : the mirror to the above, returns a plaintext string with padding removed
encryptToFile : probably the one you want to use, will write back the ciphertext, plus the initialization vector at the end to the file specified by the descriptor, please use RW permissions
genCryptoFromFile : recreates the crypto struct from an encrypted file, to be used in the decryption process

**currently the fileText + HMAC is encrypted to cipherText, then written as cipherText + IV**
please see cryptography.h and test.c for documentation and examples

changes to hashing since last time
getStringWithoutHMAC: just like the getFileContentsWithoutHMAC but accepts a string. Useful for extracting HMAC-less decrypted text where it is unsafe to write the plaintext back to a file
verifyHMACString: just like verifyHMAC but accepts a string. Useful to validate the decrypted text is authentic.

To modify an encrypted file, and re-write, it is best to re-encrypt the new plaintext and apply a new HMAC like it was a new file [see example].

This will be further discussed in our next meeting, but if you have a question please reach out.